% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAWET-package.R
\docType{package}
\name{CAWET-package}
\alias{CAWET}
\alias{CAWET-package}
\title{CAWET: Calculation of Agricultural Water Needs and Evapotranspiration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R tool for calculating theoretical irrigation water consumption over a territory. CAWET extracts, models and analyzes agro-climatic data to produce maps, graphics and comparative outputs across different scenarios. The workflow includes data extraction from online databases (via RADIS), modeling (primarily using CropWat), and visualization of results through graphics and maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inrae.r-universe.dev/CAWET}
  \item \url{https://umr-g-eau.pages-forge.inrae.fr/cawet}
}

}
\author{
\strong{Maintainer}: Pierre Rouault \email{CAWET@chaire-eacc.fr}

Authors:
\itemize{
  \item David Dorchies \email{david.dorchies@inrae.fr}
}

}
\keyword{internal}
