% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{creer_parcelles_autour_point_repgraph}
\alias{creer_parcelles_autour_point_repgraph}
\title{Create plots around a center, and change the center for each crop (for "RADIS noir" option)}
\usage{
creer_parcelles_autour_point_repgraph(
  lon,
  lat,
  crops,
  areas,
  col_crop,
  col_area
)
}
\arguments{
\item{lon}{\link{numeric} longitude in decimal degrees.}

\item{lat}{\link{numeric} latitude in decimal degrees.}

\item{crops}{\link{character} \link{vector} crops to create.}

\item{areas}{\link{numeric} \link{vector} surface areas of the crops listed in \code{crops} in m2 (in the same order as \code{crops}).}

\item{col_crop}{\link{character} standardized name of the column with the crop names.}

\item{col_area}{\link{character} standardized name of the column with the surface area.}
}
\value{
A \link[sf:sf]{sf::sf} object containing a plot for each crop, with the corresponding surface area.
}
\description{
This function creates plots around a point, with a given surface area.
}
