% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FpCAWET.R
\name{FpCAWET}
\alias{FpCAWET}
\title{Construct Full Path for CAWET}
\usage{
FpCAWET(Working_path, path)
}
\arguments{
\item{Working_path}{A string representing the working directory path.}

\item{path}{A string representing the file path to be checked or combined.}
}
\value{
A string representing the full file path.
}
\description{
This function constructs a full file path for CAWET by checking if the provided path is already an absolute path. If it is not, it combines it with the specified working directory.
}
