% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_depth.R
\name{get_soil_depth}
\alias{get_soil_depth}
\title{Get soil depth}
\usage{
get_soil_depth(sf, source = "BDGSF")
}
\arguments{
\item{sf}{An sf object containing parcels for which soil depth information is to be retrieved.
The sf object should have a column named "id_parcel" that uniquely identifies
each parcel and a column named "surface_m2" representing the surface area of
each parcel.}

\item{source}{A character string specifying the source of soil depth data. Default is "BDGSF".}
}
\value{
A data frame with parcel identifiers and their corresponding soil depth information.
}
\description{
This function retrieves soil depth information for the given parcels with a
specific grid (maille_y).
It performs a spatial intersection between the parcels and the soil depth
data obtained from the specified source, then calculates the total surface
area for each soil depth category within each parcel.
}
