% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_textures.R
\name{get_soil_textures}
\alias{get_soil_textures}
\title{Get soil textures}
\usage{
get_soil_textures(parcelles_avec_maille_y, Depth)
}
\arguments{
\item{parcelles_avec_maille_y}{An sf object containing parcels with a specific
grid (maille_y) for which soil texture information is to be retrieved. The sf
object should have a column named "id_parcel" that uniquely identifies each parcel.}

\item{Depth}{A data frame containing soil depth information for the parcels,
with a column named "id_parcel" that uniquely identifies each parcel and a
column named "soil_depth" representing the soil depth category.}
}
\value{
A data frame with parcel identifiers and their corresponding mean
values of clay, silt, and sand.
}
\description{
This function retrieves soil texture information for the given parcels with a
specific grid (maille_y).
It performs a spatial intersection between the parcels and the soil texture
data obtained from the specified source, then calculates the mean values of
clay, silt, and sand for each parcel.
}
