% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave}
\alias{ggsave}
\title{Suppress messages from ggplot2::ggsave}
\usage{
ggsave(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::ggsave}.}
}
\value{
The return value of \code{ggplot2::ggsave}, invisibly.
}
\description{
This function is a wrapper around \code{ggplot2::ggsave} that suppresses messages.
It can be used to avoid cluttering the console with messages when saving plots.
}
