% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille_safran.R
\docType{data}
\name{grille_safran}
\alias{grille_safran}
\title{SAFRAN climate grid used by CAWET}
\format{
An \code{sf} object with 8,969 rows and 4 variables:
\describe{
\item{id_maille}{Character. SAFRAN grid cell identifier.}
\item{lambx}{Integer. SAFRAN X coordinate in Lambert-93 (hectometers).}
\item{lamby}{Integer. SAFRAN Y coordinate in Lambert-93 (hectometers).}
\item{geometry}{POLYGON geometry (CRS: EPSG 2154, RGF93 / Lambert-93).}
}
}
\source{
Météo-France public dataset "coordonnees-des-mailles_339.csv":
\url{https://donneespubliques.meteofrance.fr/client/document/coordonnees-des-mailles_339.csv}
}
\usage{
grille_safran
}
\description{
\code{grille_safran} is an \code{sf} polygon grid derived from Météo-France SAFRAN
mesh coordinates. It is used to associate each territory/plot with a SAFRAN
climate cell for weather data retrieval.
}
\examples{
data(grille_safran)
sf::st_crs(grille_safran)
}
\keyword{datasets}
