% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_territory.R
\name{prepare_territory_for_existing_plots}
\alias{prepare_territory_for_existing_plots}
\title{Prepare the shapefile with the polygons for which data will be retrieved.}
\usage{
prepare_territory_for_existing_plots(
  Working_path,
  Scenario,
  COL = list(poly = "id_poly", plot = "id_parcel", crop = "code_cultu", area =
    "surface_m2"),
  col_map,
  external_plots
)
}
\arguments{
\item{Working_path}{\link{character} working directory.}

\item{Scenario}{\link{vector} contains information regarding the scenario of interest.}

\item{COL}{\link{list} standardized column names for polygon, plot, crop, and area identifiers.}

\item{col_map}{\link{data.frame} correspondence table between standardized column names (\code{col_res}) and original data column names (\code{col_user}).}

\item{external_plots}{\link{logical} If TRUE, agricultural plots are provided by the user.}
}
\value{
A \link[sf:sf]{sf::sf} object containing polygons for which data must be retrieved.
}
\description{
This function creates a sf object with polygons associated to identifiers, area, and standardized column names.
Polygons can be either agricultural plots associated to a crop and a bigger polygon, or directly the big polygons
(in which case, agricultural plots will be retrieved for each year from the RPG database).
}
