% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_territory.R
\name{prepare_territory_for_non_existing_plots}
\alias{prepare_territory_for_non_existing_plots}
\title{Prepare the shapefile with the polygons for which data will be retrieved in the case agricultural plots are fictional.}
\usage{
prepare_territory_for_non_existing_plots(
  Working_path,
  shp_link,
  COL = list(poly = "id_poly", plot = "id_parcel", crop = "code_cultu", area =
    "surface_m2"),
  col_map
)
}
\arguments{
\item{Working_path}{\link{character} working directory.}

\item{shp_link}{\link{character} name of the original data file.}

\item{COL}{\link{list} standardized column names for polygon, plot, crop, and area identifiers.}

\item{col_map}{\link{data.frame} correspondence table between standardized column names (\code{col_res}) and original data column names (\code{col_user}).}
}
\value{
A \link[sf:sf]{sf::sf} object containing a polygons for which data must be retrieved.
}
\description{
This function creates a sf object with polygons associated to identifiers, area, and standardized column names.
Polygons are agricultural plots associated to a crop and a bigger polygon
}
