% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_soil_characteristic.R
\name{process_soil_characteristic}
\alias{process_soil_characteristic}
\title{Computes the value of a soil characteristic for all relevant soil horizons, for each row of a data.frame.}
\usage{
process_soil_characteristic(df, variable, FUN)
}
\arguments{
\item{df}{\link{data.frame} result of \link[RADIS:get_soil_texture]{RADIS::get_soil_texture} or
\link[RADIS:get_soil_awc]{RADIS::get_soil_awc} joined with the result of \link[RADIS:get_soil_depth]{RADIS::get_soil_depth}
(depth in [m]).}

\item{variable}{\link{character} one of the column returned by get_soil_texture}

\item{FUN}{\link{function} the function to use to aggregate the soil
characteristic (mean or sum)}
}
\value{
A \link{numeric} vector containing averaged texture characteristic.
}
\description{
This function returns a numeric vector with a soil characteristic averaged or
summed for all soil horizons, depending on the soil depth.
}
