% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run the complete CAWET workflow}
\usage{
run(
  chemin_ordonnanceur = file.choose(),
  Scenario = read_ordonnanceur(chemin_ordonnanceur),
  Working_path = dirname(dirname(chemin_ordonnanceur)),
  cfgRun = get_Run_Config(Working_path = Working_path, Scenario = Scenario)
)
}
\arguments{
\item{chemin_ordonnanceur}{(Optional) Character string specifying the path
to the scheduler (ordonnanceur) file.
Uses an interactive file chooser (\code{\link[=file.choose]{file.choose()}}) by default.}

\item{Scenario}{(Optional) \link{data.frame} produced by \code{\link[=read_ordonnanceur]{read_ordonnanceur()}}.
If provided it is reused and passed to \code{\link[=get_Run_Config]{get_Run_Config()}} to avoid
re-reading the ordonnanceur.
By default the function calls \code{\link[=read_ordonnanceur]{read_ordonnanceur()}}.}

\item{Working_path}{(Optional) Character string specifying the working directory path.}

\item{cfgRun}{(Optional) Running configuration created by \code{\link[=get_Run_Config]{get_Run_Config()}}.}
}
\value{
Returns invisibly the running configuration (See \code{\link[=get_Run_Config]{get_Run_Config()}}).
}
\description{
This function orchestrates the entire CAWET workflow, including verification of the scheduler,
extraction of relevant information, modeling/simulation of crops, and generation of result illustrations.
}
