% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Cropwat_CAWET.R
\name{Run_Cropwat_CAWET}
\alias{Run_Cropwat_CAWET}
\title{Run Cropwat CAWET simulation for a given crop, irrigation method and sequence}
\usage{
Run_Cropwat_CAWET(
  Irrigation_param,
  mod_irr,
  Sequence,
  carac_date_irr,
  Link_RPG_mod,
  Crop,
  Crops_param,
  Meteo_maille,
  liste_not_parametredcrops
)
}
\arguments{
\item{Irrigation_param}{A data frame containing irrigation parameters.}

\item{mod_irr}{A string representing the irrigation method.}

\item{Sequence}{A data frame containing sequence information.}

\item{carac_date_irr}{A logical value indicating whether to use specific irrigation dates}

\item{Link_RPG_mod}{A data frame linking RPG crop names to Cropwat crop names.}

\item{Crop}{A string representing the crop name.}

\item{Crops_param}{A data frame containing crop parameters.}

\item{Meteo_maille}{A data frame containing meteorological data for the grid cell}

\item{liste_not_parametredcrops}{A vector of strings representing crops that are
not parameterized.}
}
\value{
None. The function saves the detailed output results to a CSV file.
}
\description{
This function runs the Cropwat CAWET simulation for a specified crop, irrigation method, and sequence. It takes into account various parameters such as irrigation settings, crop characteristics, meteorological data, and simulation settings to generate detailed output results.
}
