% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.1-Graphiques_simulation_scenario_SIMCAWET.R
\name{run_script3_1}
\alias{run_script3_1}
\title{Plotting simulation results for a CAWET scenario}
\usage{
run_script3_1(cfgRun = get_Run_Config(Working_path = Working_path))
}
\arguments{
\item{cfgRun}{(Optional) Running configuration created by \code{\link[=get_Run_Config]{get_Run_Config()}}.}
}
\value{
This function does not return a value but saves graphs to the specified directory.
}
\description{
This script generates graphs based on the results of a CAWET scenario simulation.
It reads the simulation results and creates various visualizations such as pie charts for crop distribution,
bar charts for annual and monthly irrigation data, and comparison graphs with validation data if provided.
It saves the generated graphs in a specified directory for further analysis.
}
