% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.2-Comparatif_scenarii.R
\name{run_script3_2}
\alias{run_script3_2}
\title{Scenario Comparison Script}
\usage{
run_script3_2(cfgRun = get_Run_Config(Working_path = Working_path))
}
\arguments{
\item{cfgRun}{(Optional) Running configuration created by \code{\link[=get_Run_Config]{get_Run_Config()}}.}
}
\value{
This function does not return a value but generates comparison plots and CSV files in the Results_comparison folder.
}
\description{
This function runs the script 3.2 for comparing scenarios based on the last run or a chosen run path.
It generates comparison plots and CSV files for area, irrigation, and ETc in the Results_comparison folder.
}
