% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.3-Indicateurs_climatiques.R
\name{run_script3_3}
\alias{run_script3_3}
\title{Climatic indicators calculation and graphs generation}
\usage{
run_script3_3(cfgRun = get_Run_Config(Working_path = Working_path))
}
\arguments{
\item{cfgRun}{(Optional) Running configuration created by \code{\link[=get_Run_Config]{get_Run_Config()}}.}
}
\value{
None. The function saves the results and graphs in the specified directories.
}
\description{
This function calculates climatic indicators from meteorological data and generates corresponding graphs.
}
