% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_CAWET.R
\name{setup_CAWET}
\alias{setup_CAWET}
\title{Setup CAWET instance in a specified directory}
\usage{
setup_CAWET(destdir)
}
\arguments{
\item{destdir}{Character string specifying the destination directory
where the CAWET instance should be set up.}
}
\value{
None. The function performs file copy operations.
}
\description{
This function copies all necessary files from the CAWET package's
extdata directory to a user-specified destination directory.
}
\examples{
destdir <- file.path(tempdir(), "CAWET")
setup_CAWET(destdir)
list.files(destdir)
}
