% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{st_read}
\alias{st_read}
\title{Suppress messages from sf::st_read}
\usage{
st_read(...)
}
\arguments{
\item{...}{Arguments passed to \code{sf::st_read}.}
}
\value{
The return value of \code{sf::st_read}, invisibly.
}
\description{
This function is a wrapper around \code{sf::st_read} that suppresses messages.
It can be used to avoid cluttering the console with messages when reading spatial data.
}
