% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verification_ordonnanceur.R
\name{Verif_ordonnanceur_CAWET}
\alias{Verif_ordonnanceur_CAWET}
\title{Ordonnanceur verification function for CAWET}
\usage{
Verif_ordonnanceur_CAWET(cfgRun)
}
\arguments{
\item{cfgRun}{(Optional) Running configuration created by \code{\link[=get_Run_Config]{get_Run_Config()}}.}
}
\value{
None. The function stops execution and raises an error if any verification fails.
}
\description{
This function verifies the content of the ordonnanceur (scenario file) for CAWET. It checks for the existence and correctness of various parameters and files specified in the ordonnanceur.
}
